refactor: restructure the modal & add location page
