+++ packages/client/src/views/FavoriteLocation/FavLocation.tsx
@@ -1,0 +1,3 @@
+export const FavLocation = () => {
+  return <p>FAV LOCATION PAGE</p>;
+};
